requirements.txt
-f /usr/share/pip-wheels
django-cors-headers==3.1.0
django-filter==2.2.0
django-taggit==1.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.10.3
Markdown==3.1.1
mysqlclient==1.4.2.post1
requests==2.9.1
